+++ packages/vue-component-library/src/lib-components/MobileDrawer.vue
@@ -54,6 +54,7 @@ onMounted(() => {
       />
       <button
         class="mobile-button"
+        aria-label="Toggle Menu"
         :class="isDropdownExpandedClass"
         @click="handleDropdownExpansion"
       >
@@ -77,11 +78,15 @@ onMounted(() => {
         v-if="isDropdownExpanded"
         class="button-dropdown-modal"
       >
-        <SvgGlyphClose
+        <button
           v-if="isMobile"
           class="svg-glyph-close"
+          aria-label="Close"
+          type="button"
           @click="removeOverlay"
-        />
+        >
+          <SvgGlyphClose />
+        </button>
         <div
           class="button-dropdown-modal-wrapper"
           :class="isDropdownExpandedClass"