+++ packages/vue-component-library/src/stories/SiteNotificationBanner.spec.js
@@ -1,0 +1,11 @@
+describe('GLOBAL / SiteNotificationBanner', () => {
+  it('Default', () => {
+    cy.visit(
+      '/iframe.html?id=global-site-notification-banner--default&args=&viewMode=story'
+    )
+
+    cy.get('.site-notification-banner').should('exist')
+
+    cy.percySnapshot('GLOBAL / SiteNotificationBanner: Default')
+  })
+})